refactor: type hints + avoid mutable defaults
