+++ cognite/neat/_graph/extractors/_classic_cdf/_classic.py
@@ -7,7 +7,7 @@ from rdflib import Namespace
 
 from cognite.neat._constants import DEFAULT_NAMESPACE
 from cognite.neat._graph.extractors._base import BaseExtractor
-from cognite.neat._graph.models import Triple
+from cognite.neat._shared import Triple
 from cognite.neat._utils.collection_ import chunker
 from cognite.neat._utils.rdf_ import remove_namespace_from_uri
