+++ cognite/neat/_graph/extractors/_dexpi.py
@@ -8,7 +8,7 @@ from rdflib import RDF, RDFS, XSD, Literal, Namespace, URIRef
 
 from cognite.neat._constants import DEFAULT_NAMESPACE
 from cognite.neat._graph.extractors._base import BaseExtractor
-from cognite.neat._graph.models import Triple
+from cognite.neat._shared import Triple
 from cognite.neat._utils.rdf_ import as_neat_compliant_uri
 from cognite.neat._utils.xml_ import get_children, iterate_tree
