--- cognite/neat/_graph/models.py
@@ -1,7 +1,0 @@
-from typing import TypeAlias
-
-from rdflib import Literal
-from rdflib.term import URIRef
-
-Triple: TypeAlias = tuple[URIRef, URIRef, Literal | URIRef]
-InstanceType: TypeAlias = URIRef
