+++ cognite/neat/_graph/queries/_base.py
@@ -8,10 +8,10 @@ from rdflib import Literal as RdfLiteral
 from rdflib.query import ResultRow
 
 from cognite.neat._constants import UNKNOWN_TYPE
-from cognite.neat._graph.models import InstanceType
 from cognite.neat._rules._constants import EntityTypes
 from cognite.neat._rules.models.entities import ClassEntity
 from cognite.neat._rules.models.information import InformationRules
+from cognite.neat._shared import InstanceType
 from cognite.neat._utils.rdf_ import remove_namespace_from_uri
 
 from ._construct import build_construct_query
